// apps/client/src/types/types.ts
export type ContainerSize = Database["public"]["Enums"]["container_size"];
// ...
export interface Order {
  fullName: string;
  itemName: string;
  container: ContainerSize;
  weight: number;
  flavor: string;
  flavorLabel: string;
  protein: string;
  proteinLabel: string;
  quantity: number;
}
export interface OrderError {
  error: Error | null;
  message: string;
  order: Order;
}
export interface IngredientAmounts {
  [name: string]: {
    label: string;
    amount: number;
    lbsPer: number;
    units: string;
  };
}
export interface Meal {
  protein: string;
  proteinLabel: string;
  flavor: string;
  flavorLabel: string;
  originalWeight: number;
  weight: number;
  weightLbOz: string;
  backstockWeight: number;
  cookedWeight: number;
  displayColor: string | null;
}
// ...
export type AllBackstockRow = ToCamelCase<SupaAllBackstockRow>;
export type FlavorRow = ToCamelCase<SupaFlavorRow>;
export type OrderHeaderRow = ToCamelCase<SupaOrderHeaderRow>;
export type ProteinRow = ToCamelCase<SupaProteinRow>;
// ...
export interface ProteinWeights {
  [protein: string]: {
    [flavor: string]: {
      proteinLabel: string;
      flavorLabel: string;
      weight: number;
    };
  };
}